+++ src/ScreenNetEvaluation.cpp
@@ -187,7 +187,7 @@ void ScreenNetEvaluation::UpdateStats()
 	if( THEME->GetMetricB(m_sName,"ShowGradeArea") )
 		m_Grades[m_pActivePlayer].SetGrade( (Grade)NSMAN->m_EvalPlayerData[m_iCurrentPlayer].grade );
 	if( THEME->GetMetricB(m_sName,"ShowScoreArea") )
-		m_textScore[m_pActivePlayer].SetTargetNumber( NSMAN->m_EvalPlayerData[m_iCurrentPlayer].score );
+		m_textScore[m_pActivePlayer].SetTargetNumber( static_cast<float>(NSMAN->m_EvalPlayerData[m_iCurrentPlayer].score) );
 
 	//Values greater than 6 will cause a crash
 	if( NSMAN->m_EvalPlayerData[m_iCurrentPlayer].difficulty < 6 )
@@ -202,7 +202,7 @@ void ScreenNetEvaluation::UpdateStats()
 		// should not be shown.
 		if( !m_textJudgmentLineNumber[j][m_pActivePlayer].GetName().empty() )
 		{
-			m_textJudgmentLineNumber[j][m_pActivePlayer].SetTargetNumber( NSMAN->m_EvalPlayerData[m_iCurrentPlayer].tapScores[j] );
+			m_textJudgmentLineNumber[j][m_pActivePlayer].SetTargetNumber( static_cast<float>(NSMAN->m_EvalPlayerData[m_iCurrentPlayer].tapScores[j]) );
 		}
 	}
